uppercase tags
just noticed that we're supposed to use middleware to uppercase our tags. I have no idea where to even start on this but I tried adding something to one of my requests that ought to work, even though it's not middleware. Also commented my code to try and make it less crazy to read, but until I figure out how to break it into separate files it's just going to be ridiculously long and there's not much I can do about that. Shifting gears to work on react stuff because the server stuff is making me crazy right now.